fix generated default cert, update client verifier
We were previously disabling all cert and tls
signature verification in order to get a connection
to the server working. The errors were caused by
the server exposing a certificate that the rustls
client's safe defaults consider unsafe. Adding
a custom openssl config and updating the cert
generation command lets us re-enable tls signature
verification. The NoVerifyVerifier is replaced
with an ExactCertVerifier to assert that the
server and client share a certificate.